+++ src/parser/inline/autolink/email.test.ts
@@ -9,12 +9,15 @@ describe('Unit: parser/inline/autolink/email', () => {
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('a@')), [['a@'], '']);
-      assert.deepStrictEqual(inspect(parser('a@_')), [['a@'], '_']);
-      assert.deepStrictEqual(inspect(parser('a@-')), [['a@'], '-']);
-      assert.deepStrictEqual(inspect(parser('a@_b')), [['a@'], '_b']);
-      assert.deepStrictEqual(inspect(parser('a@-b')), [['a@'], '-b']);
+      assert.deepStrictEqual(inspect(parser('a@_')), [['a@_'], '']);
+      assert.deepStrictEqual(inspect(parser('a@-')), [['a@-'], '']);
+      assert.deepStrictEqual(inspect(parser('a@_b')), [['a@_b'], '']);
+      assert.deepStrictEqual(inspect(parser('a@-b')), [['a@-b'], '']);
+      assert.deepStrictEqual(inspect(parser('a@b@')), [['a@b@'], '']);
+      assert.deepStrictEqual(inspect(parser('a@bc@')), [['a@bc@'], '']);
+      assert.deepStrictEqual(inspect(parser('a@b@c')), [['a@b@c'], '']);
       assert.deepStrictEqual(inspect(parser('a@@')), [['a@@'], '']);
-      assert.deepStrictEqual(inspect(parser('a@@b')), [['a@@'], 'b']);
+      assert.deepStrictEqual(inspect(parser('a@@b')), [['a@@b'], '']);
       assert.deepStrictEqual(inspect(parser(' a@b')), undefined);
     });
 
@@ -29,8 +32,6 @@ describe('Unit: parser/inline/autolink/email', () => {
       assert.deepStrictEqual(inspect(parser('a@a--b')), [['<a class="email" href="mailto:a@a--b" rel="noopener">a@a--b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('a@http://host')), [['<a class="email" href="mailto:a@http" rel="noopener">a@http</a>'], '://host']);
       assert.deepStrictEqual(inspect(parser('a@ttp://host')), [['<a class="email" href="mailto:a@ttp" rel="noopener">a@ttp</a>'], '://host']);
-      assert.deepStrictEqual(inspect(parser('a@b@')), [['<a class="email" href="mailto:a@b" rel="noopener">a@b</a>'], '@']);
-      assert.deepStrictEqual(inspect(parser('a@b@c')), [['<a class="email" href="mailto:a@b" rel="noopener">a@b</a>'], '@c']);
       assert.deepStrictEqual(inspect(parser('a@a#')), [['<a class="email" href="mailto:a@a" rel="noopener">a@a</a>'], '#']);
       assert.deepStrictEqual(inspect(parser('a@a#1')), [['<a class="email" href="mailto:a@a" rel="noopener">a@a</a>'], '#1']);
     });